Move PriceEstimations to it's own component
