+++ src/components/trade/PriceEstimations/PriceEstimations.tsx
